feat(cppast): add support for identifying namespaces
This commit adds support for identifying namespaces in the CPPBasicIdentListenerTest. A new test case `shouldIdentifyNameSpace` is added to test the identification of namespaces. The `CPPBasicIdentListener` is modified to handle the `enterNamespaceDefinition` event and extract the namespace name from the context. The extracted namespace name is then assigned to the `PackageName` property of the `CodeContainer` class.