feat(cpp): add support for identifying C++ structures #24
This commit adds support for identifying C++ structures in the `CPPBasicIdentListener` class. It includes a new test case that analyzes a C++ code snippet and verifies the identification of a structure. The identified structure is then added to the `CodeContainer` object. Additionally, the `CodeDataStruct` class is updated to include a new `DataStructType` enum value for C++ unions.