feat(c): add support for preprocessing C code #24
This commit adds support for preprocessing C code using the `org.anarres.cpp` library. It introduces a new `CAnalyser` class that initializes a `Preprocessor` object and adds features and warnings. The `addSource` method is added to add source code to the program. The `analysis` method now preprocesses the code using the `Preprocessor` object and then parses the resulting output. The `CParser.g4` file is modified to include support for type keywords and pointers. The `CFullIdentListenerTest.kt` file is modified to update the test cases. The `build.gradle.kts` file is modified to include the `org.anarres:jcpp` dependency.